他のインターフェース名に合わせてプラットフォームAPI接続クラスのインターフェース名を変更した。
他のクラス名に合わせてプラットフォームAPI接続クラス名を変更した。
